feat(notifications): renaming service
